20240701 - Fix the polars version bug in CST.py 20240703 - Fix the codon range bugs of MetaCodon.py.

+++ scripts/foo/CST.py
@@ -378,8 +378,12 @@ class CodonSelectiveTime(object):
 
         if initiate_rate:
             high_rna = self.high_rna.with_columns(
+                # self.high_rna[rna_sample] * self.high_rna['name'].apply(lambda x: self.gene_table[rpf_sample][x].initiate_rate if x in self.gene_table[rpf_sample].keys() else 1))
+                # Calling `map_elements` without specifying `return_dtype` can lead to unpredictable results. Specify `return_dtype` to silence this warning
                 self.high_rna[rna_sample] * self.high_rna['name'].apply(
-                    lambda x: self.gene_table[rpf_sample][x].initiate_rate if x in self.gene_table[rpf_sample].keys() else 1))
+                    lambda x: self.gene_table[rpf_sample][x].initiate_rate if x in self.gene_table[rpf_sample].keys() else 1,
+                    return_dtype=pl.Float32))
+            
         else:
             high_rna = self.high_rna
